requirements.txt
Flask
Pillow
Flask-CORS